+++ NEWS.md
@@ -1,7 +1,8 @@
 # teal.modules.clinical 0.8.3.9000
 
 * Extend `tm_t_coxreg` to optionally produce univariate Cox regressions.
-* Updated `tm_t_binary_outcome` to display Odds Ratio estimates and include new methods for CIs and p-values.
+* Updated `tm_t_binary_outcome` to display Odds Ratio estimates, include new methods for CIs and p-values and 
+display a summary for individual response categories.
 
 # teal.modules.clinical 0.8.3