+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.modules.clinical 0.8.3.9000
+# teal.modules.clinical 0.8.4
 
 * Extend `tm_t_coxreg` to optionally produce univariate Cox regressions.
 * Updated `tm_t_binary_outcome` to display Odds Ratio estimates, include new methods for CIs and p-values and